+++ backend/src/nodes/nodes/pytorch/pytorch_alt_nodes.py
@@ -13,14 +13,14 @@ from facexlib.utils.face_restoration_helper import FaceRestoreHelper
 
 from appdirs import user_data_dir
 
-from .categories import PyTorchCategory
-from .node_base import NodeBase
-from .node_factory import NodeFactory
-from .properties.inputs import *
-from .properties.outputs import *
-from .utils.utils import get_h_w_c, np2tensor, tensor2np
-from .utils.exec_options import get_execution_options
-from .utils.torch_types import PyTorchModel
+from ...categories import PyTorchCategory
+from ...node_base import NodeBase
+from ...node_factory import NodeFactory
+from ...properties.inputs import *
+from ...properties.outputs import *
+from ...utils.utils import get_h_w_c, np2tensor, tensor2np
+from ...utils.exec_options import get_execution_options
+from ...utils.torch_types import PyTorchModel
 from .pytorch_nodes import to_pytorch_execution_options
