+++ backend/src/run.py
@@ -7,6 +7,8 @@ import sys
 import traceback
 from json import dumps as stringify
 from typing import Any, Dict, List, Optional, TypedDict
+import importlib
+import os
 
 # pylint: disable=unused-import
 import cv2
@@ -17,57 +19,9 @@ from sanic.response import json
 from sanic_cors import CORS
 
 from nodes.categories import categories, category_order
-
-# pylint: disable=ungrouped-imports,wrong-import-position
-from nodes import image_adj_nodes  # type: ignore
-from nodes import image_chan_nodes  # type: ignore
-from nodes import image_dim_nodes  # type: ignore
-from nodes import image_filter_nodes  # type: ignore
-from nodes import image_iterator_nodes  # type: ignore
-from nodes import image_nodes  # type: ignore
-from nodes import image_util_nodes  # type: ignore
-
-try:
-    import torch
-
-    # Lazily initialize to avoid unnecessary ram usage
-    if torch.cuda.is_available():
-        # pylint: disable=W0212
-        torch.cuda._lazy_init()
-
-    # pylint: disable=unused-import,ungrouped-imports
-    from nodes import pytorch_nodes  # type: ignore
-    from nodes import pytorch_iterators  # type: ignore
-    from nodes import pytorch_alt_nodes  # type: ignore
-except Exception as e:
-    torch = None
-    logger.warning(e)
-    logger.info("PyTorch most likely not installed")
-
-try:
-    import onnx
-    import onnxruntime
-
-    # pylint: disable=unused-import,ungrouped-imports
-    from nodes import onnx_nodes  # type: ignore
-except Exception as e:
-    logger.warning(e)
-    logger.info("ONNX most likely not installed")
-
-
-try:
-    # pylint: disable=unused-import
-    import ncnn_vulkan
-
-    # pylint: disable=unused-import,ungrouped-imports
-    from nodes import ncnn_nodes  # type: ignore
-except Exception as e:
-    logger.warning(e)
-    logger.info("NCNN most likely not installed")
-
-# pylint: disable=unused-import
-from nodes import utility_nodes  # type: ignore
 from nodes.node_factory import NodeFactory
+from nodes.utils.exec_options import set_execution_options, ExecutionOptions
+
 from base_types import NodeId, InputId, OutputId
 from chain.cache import OutputCache
 from chain.json import parse_json, JsonNode
@@ -81,7 +35,24 @@ from response import (
     noExecutorResponse,
     successResponse,
 )
-from nodes.utils.exec_options import set_execution_options, ExecutionOptions
+
+
+# Dynamically import all nodes
+for root, dirs, files in os.walk(
+    os.path.join(os.path.dirname(__file__), "nodes", "nodes")
+):
+    for file in files:
+        if file.endswith(".py") and not file.startswith("_"):
+            module = os.path.relpath(
+                os.path.join(root, file), os.path.dirname(__file__)
+            )
+            module = module.replace(os.path.sep, ".")[:-3]
+            try:
+                importlib.import_module(f"{module}", package=None)
+            except ImportError as e:
+                logger.warning(f"Failed to import {module}: {e}")
+                if "torch" in str(e):
+                    torch = None
 
 
 class AppContext:
